Configure Tyrus client from system properties
We use only javax.websocket classes by default but
if the there are `System.getProperties()` starting with
`org.glassfish.tyrus.client.` then we create explicitly
a Tyrus client and set the properties on it. This enables
the user to e.g. configure a HTTP proxy or connection
keep-alive without passing a custom client instance or any
code change.